+++ R/standardise_sumstats_column_headers_crossplatform.R
@@ -1,26 +1,43 @@
 #' Standardise the column headers in the Summary Statistics files
 #'
 #' Use a reference data table of common column header names (stored in
-#' sumstatsColHeaders or user inputted mapping file) to convert them to a
+#' \code{sumstatsColHeaders} or user inputted mapping file) to convert them to a
 #' standard set, i.e. chromosome -> CHR. This function does not check that all
 #' the required column headers are present. The amended header is written
 #' directly back into the file
 #'
-#' @inheritParams format_sumstats
+#' @param uppercase_unmapped For columns that could not be identified in 
+#' the \code{mapping_file}, return them in the same format they were input as
+#'  (without forcing them to uppercase). 
+#' @inheritParams format_sumstats 
+#' @inheritParams compute_nsize
 #' @return list containing sumstats_dt, the modified summary statistics data
 #' table object
-#' @keywords internal
+#' @export
 #' @importFrom data.table setnames
+#' @examples
+#' sumstats_dt <- data.table::fread(system.file("extdata", "eduAttainOkbay.txt",
+#'                                              package = "MungeSumstats"))
+#' sumstats_dt2 <- standardise_sumstats_column_headers_crossplatform(
+#'     sumstats_dt=sumstats_dt)
 standardise_sumstats_column_headers_crossplatform <-
     function(sumstats_dt,
-             mapping_file = sumstatsColHeaders) {
+             mapping_file = sumstatsColHeaders,
+             uppercase_unmapped=TRUE,
+             return_list=TRUE) {
+          
         message("Standardising column headers.")
         message("First line of summary statistics file: ")
         msg <- paste0(names(sumstats_dt), split = "\t")
         message(msg)
-        # first make all column headers upper case
-        data.table::setnames(sumstats_dt, toupper(names(sumstats_dt)))
-        column_headers <- names(sumstats_dt)
+        #### Store original colnames #####
+        ## In case we want to use the original casing 
+        ## IMPORTANT! Must use copy() function or else this vector will get 
+        ## changed when the data.table it comes from gets changed.
+        names_og <- data.table::copy(names(sumstats_dt))
+        #### first make all column headers upper case ####
+        data.table::setnames(sumstats_dt, toupper(names(sumstats_dt))) 
+        column_headers <- names(sumstats_dt) 
         # load synonym mapping - internal data no loading
         # Go through each and get correct spelling
         # allow for differing cases column names
@@ -46,5 +63,23 @@ standardise_sumstats_column_headers_crossplatform <-
             data.table::setnames(sumstats_dt, "A1", "A2", skip_absent = TRUE)
             data.table::setnames(sumstats_dt, "A*", "A1")
         }
-        return(list("sumstats_dt" = sumstats_dt))
+        #### Return mapped columns to original casing if requested ####
+        if(!uppercase_unmapped){
+            unmapped <- colnames(sumstats_dt)[
+                !colnames(sumstats_dt) %in% mapping_file$CORRECTED
+            ]
+            unmapped_og <- names_og[toupper(names_og) %in% unmapped]
+            if(length(unmapped_og)>0){
+                msg <- paste("Returning unmapped column names",
+                             "without making them uppercase.")
+                message(msg)
+                data.table::setnames(sumstats_dt, unmapped, unmapped_og)
+            } 
+        }
+        #### Return format ####
+        if(return_list){
+            return(list("sumstats_dt" = sumstats_dt))
+        }else {
+            return(sumstats_dt)
+        } 
     }